% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edit_ORYZA_file.R
\name{edit_ORYZA_file}
\alias{edit_ORYZA_file}
\title{edit Oryza file}
\usage{
edit_ORYZA_file(variables, values, file_path)
}
\arguments{
\item{variables}{name of variable in Oryza file}

\item{values}{values to be written to}

\item{file_path}{the file path of the file to be edited}
}
\description{
Edit Oryza file
}
\details{
This function finds variables specified and re-writes them
   with new values as specified. Multiple values can be changed at once.
   Warning: This function alters the Oryza file.
}
\author{
Matthew Espe
}

