% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_write_ORYZA.R
\name{find_phen_pred}
\alias{find_phen_pred}
\title{Find phenology predictions}
\usage{
find_phen_pred(reruns, DAP = FALSE)
}
\arguments{
\item{reruns}{a \code{list} object, each element corresponding to
rerun - produced by \code{split_res}}

\item{DAP}{logical - should results be returned as day of year
    or as DAP (days after planting)}
}
\value{
a \code{matrix} with 3 columns: prediced panicle initiation,
    50% heading, and physiological maturity.  Each row corresponds
    to the a single rerun in the res.dat file.
}
\description{
Find phenology predictions from Oryza
}
\details{
Find the predicted dates of each phenological event
    reruns = split reruns file produced by split_res()
    DAP = days after planting(simulation start) otherwise gives day of year
}
\author{
Matthew Espe
}

