% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_current.R
\name{get_current_oryza}
\alias{get_current_oryza}
\title{Get current Oryza settings}
\usage{
get_current_oryza(file_path)
}
\arguments{
\item{file_path}{complete file path to control file, including
    the name of the file to be read.}
}
\value{
active files being referred to in control file
}
\description{
Get the current files being referred to in an Oryza(v3) control file
}
\details{
This is a convienence function that reads the Oryza(v3)
   control file specified and prints the current active files
   to the console.
}
\author{
Matthew Espe
}

