% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_write_ORYZA.R
\name{get_pred_mat}
\alias{get_pred_mat}
\title{Get matrix of Oryza(v3) predictions}
\usage{
get_pred_mat(file_path, DAP = TRUE)
}
\arguments{
\item{file_path}{character string of file-path to location where
res.dat and op.dat are currently saved.}

\item{DAP}{logical - should results be return as day of year, or
    DAP (days after planting).}
}
\value{
a \code{matrix} with 4 columns: prediced panicle initiation,
    50% heading, physiological maturity, and yield.  Each row corresponds
    to the a single rerun in the res.dat file.
}
\description{
Get prediciton matrix
}
\details{
Function to read model output and return predictions.
    This discards the first run (run of .exp file).
}
\author{
Matthew Espe
}

