% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POWER2ORYZA.R
\name{getPOWER}
\alias{getPOWER}
\title{getPOWER}
\usage{
getPOWER(lat, lon, ms = 1, ds = 1, ys, me = 12, de = 31, ye)
}
\arguments{
\item{lat}{latitude}

\item{lon}{longitude}

\item{ms}{start month}

\item{ds}{start day}

\item{ys}{start year}

\item{me}{end month}

\item{de}{end day}

\item{ye}{end year}
}
\value{
\code{data.frame} object with POWER data
}
\description{
Import data from NASA POWER into R
}
\details{
This function imports data from the NASA POWER database
    at http://power.larc.nasa.gov/cgi-bin/cgiwrap/solar/agro.cgi.
}
\author{
Matthew Espe
}

