% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/POWER2ORYZA.R
\name{POWER2ORYZA}
\alias{POWER2ORYZA}
\title{POWER2ORYZA}
\usage{
POWER2ORYZA(year, station_nbr, lat, long, prefix)
}
\arguments{
\item{year}{year fo query}

\item{station_nbr}{user defined value for the station number
to be used for the name of the Oryza weather file.}

\item{lat}{latitude}

\item{long}{longitude}

\item{prefix}{prefix to be used for saved file}
}
\description{
Retrieve and process NASA POWER data for Oryza(v3)
}
\details{
This function retrieves data from NASA POWER for an
    entire year, and processes it for use by Oryza(v3).
    Weather file is saved to CURRENT WORKING DIRECTORY.
}
\author{
Matthew Espe
}

