% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_write_ORYZA.R
\name{read_res}
\alias{read_res}
\title{Read res.dat}
\usage{
read_res()
}
\value{
a single \code{data.frame} object with all data rows
    from the res.dat file.
}
\description{
Read in res.dat file
}
\details{
This function uses cygwin commands to pre-process the res.dat file
    that is produced by Oryza(v3). Will not work if current working
    directory does not contain the res.dat file or if output is
    named differently than res.dat.
}
\author{
Matthew Espe
}

