% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_write_ORYZA.R
\name{split_res}
\alias{split_res}
\title{split res}
\usage{
split_res(res_dataframe)
}
\arguments{
\item{res_dataframe}{a \code{data.frame} containing the data from the res.dat
    file - created by \code{read_res}}
}
\value{
a \code{list} object with each element corresponding to a rerun
    from Oryza(v3)
}
\description{
Split into reruns
}
\details{
Splits the res dataframe into a list with each rerun as an element.
}
\author{
Matthew Espe
}

