\name{Roryza-package}

\alias{Roryza-package}

\alias{Roryza}

\docType{package}

\title{
\packageTitle{Roryza}
}

\description{
This package provides an R interface to a C++ implementation of the ORYZA simulation model of rice growth and development.
}


\references{
Bouman, B.A.M., Kropff, M.J., Tuong, T.P., Wopereis, M.C.S., ten Berge, H.F.M., & van Laar, H.H. (2001). ORYZA2000 : modeling lowland rice. International Rice Research Institute / Wageningen University. https://books.irri.org/9712201716_content.pdf
}
